+++ src/snowplow_signals/models/__init__.py
@@ -2,19 +2,18 @@ from .model import AttributeInput as Attribute
 from .model import (
     AttributeOutput,
     BatchSource,
-    AttributeInput as Attribute,
-    FieldModel as Field,
-    Event,
-    Entity,
     Criteria,
     Criterion,
+    Entity,
+    Event,
     LinkEntity,
-    GetOnlineAttributesRequest,
-    LinkEntity,
-    TestViewRequest,
-    ViewOutput,
 )
-
+from .model import FieldModel as Field
+from .model import (
+    GetOnlineAttributesRequest,
+    TestViewRequest,
+    ViewResponse,
+)
 from .online_attributes_response import OnlineAttributesResponse
 from .service import Service
 from .view import View
@@ -23,7 +22,7 @@ AttributeOutput
 Criteria
 Criterion
 View
-ViewOutput
+ViewResponse
 Service
 Entity
 TestViewRequest